refactor(core): introduce repository traits
Add NoteRepository, TimelineRepository, ProjectRepository, and
TaskRepository traits with filesystem implementations. Free functions
now delegate to repository instances for backward compatibility.
