+++ core/src/project/mod.rs
@@ -1,21 +1,19 @@
 mod model;
+pub mod repository;
 pub mod task;
 
 pub use model::{ProjectActivitySummary, ProjectSummary};
+pub use repository::{FsProjectRepository, ProjectRepository};
 pub use task::{
     TaskSummary, complete_task, create_task, delete_task, list_active_tasks, list_done_tasks,
     update_task,
 };
 
-use std::fs;
 use std::path::{Path, PathBuf};
 
-use chrono::{DateTime, FixedOffset, Local, NaiveDate};
+use chrono::NaiveDate;
 
 use crate::error::CoreError;
-use crate::infra::fs_helpers;
-use crate::infra::paths;
-use crate::shared::frontmatter::{self, ProjectFrontmatter};
 use crate::shared::validated::Slug;
 
 pub fn create_project(
@@ -24,72 +22,15 @@ pub fn create_project(
     name: &str,
     description: &str,
 ) -> Result<PathBuf, CoreError> {
-    let file_path = paths::project_file_path(base_dir, slug.as_str());
-    if file_path.exists() {
-        return Err(CoreError::AlreadyExists(format!("project: {slug}")));
-    }
-
-    let slug_str = slug.as_str();
-    let proj_dir = paths::project_dir(base_dir, slug_str);
-    fs::create_dir_all(paths::active_tasks_dir(base_dir, slug_str))?;
-    fs::create_dir_all(paths::done_tasks_dir(base_dir, slug_str))?;
-
-    let now: DateTime<FixedOffset> = Local::now().into();
-    let fm = ProjectFrontmatter {
-        name: name.to_string(),
-        created: now,
-        description: description.to_string(),
-    };
-    let content = frontmatter::render(&fm, "")?;
-    let file_path = paths::project_file_path(base_dir, slug_str);
-    fs::write(&file_path, content)?;
-
-    Ok(proj_dir)
+    FsProjectRepository::new(base_dir.to_path_buf()).create(slug, name, description)
 }
 
 pub fn list_projects(base_dir: &Path) -> Result<Vec<ProjectSummary>, CoreError> {
-    let dir = paths::projects_dir(base_dir);
-    if !dir.exists() {
-        return Ok(Vec::new());
-    }
-
-    let mut projects = Vec::new();
-    let mut entries: Vec<_> = fs::read_dir(&dir)?
-        .filter_map(|e| e.ok())
-        .filter(|e| e.path().is_dir())
-        .collect();
-    entries.sort_by_key(|e| e.file_name());
-
-    for entry in entries {
-        let slug_str = entry.file_name().to_string_lossy().to_string();
-        if let Ok(slug) = Slug::parse(&slug_str) {
-            if let Ok(summary) = read_project(base_dir, &slug) {
-                projects.push(summary);
-            }
-        }
-    }
-
-    Ok(projects)
+    FsProjectRepository::new(base_dir.to_path_buf()).list()
 }
 
 pub fn read_project(base_dir: &Path, slug: &Slug) -> Result<ProjectSummary, CoreError> {
-    let slug_str = slug.as_str();
-    let file_path = paths::project_file_path(base_dir, slug_str);
-    if !file_path.exists() {
-        return Err(CoreError::NotFound(format!("project: {slug}")));
-    }
-
-    let content = fs::read_to_string(&file_path)?;
-    let (fm, _body): (ProjectFrontmatter, String) = frontmatter::parse(&content)?;
-
-    let active_dir = paths::active_tasks_dir(base_dir, slug_str);
-    let active_task_count = fs_helpers::count_md_files(&active_dir)?;
-
-    Ok(ProjectSummary::from_frontmatter(
-        slug_str.to_string(),
-        fm,
-        active_task_count,
-    ))
+    FsProjectRepository::new(base_dir.to_path_buf()).read(slug)
 }
 
 pub fn get_project_activity_summary(
@@ -131,8 +72,10 @@ pub fn get_project_activity_summary(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::shared::frontmatter::TaskFrontmatter;
-    use chrono::TimeZone;
+    use crate::infra::paths;
+    use crate::shared::frontmatter::{self, TaskFrontmatter};
+    use chrono::{DateTime, FixedOffset, TimeZone};
+    use std::fs;
     use tempfile::TempDir;
 
     fn slug(s: &str) -> Slug {
